migration & language picker

+++ src/backoffice/packages/package-section/views/created/packages-created-overview.element.ts
@@ -6,6 +6,7 @@ import { UUIPaginationEvent } from '@umbraco-ui/uui';
 import { PackageDefinitionModel, PackageResource } from '@umbraco-cms/backend-api';
 import { UmbLitElement } from '@umbraco-cms/element';
 import { tryExecuteAndNotify } from '@umbraco-cms/resources';
+import { UmbModalService, UMB_MODAL_SERVICE_CONTEXT_TOKEN } from '@umbraco-cms/modal';
 
 @customElement('umb-packages-created-overview')
 export class UmbPackagesCreatedOverviewElement extends UmbLitElement {
@@ -50,6 +51,8 @@ export class UmbPackagesCreatedOverviewElement extends UmbLitElement {
 	@state()
 	private _total?: number;
 
+	private _modalService?: UmbModalService;
+
 	constructor() {
 		super();
 	}
@@ -57,6 +60,10 @@ export class UmbPackagesCreatedOverviewElement extends UmbLitElement {
 	connectedCallback(): void {
 		super.connectedCallback();
 		this.#getPackages();
+
+		this.consumeContext(UMB_MODAL_SERVICE_CONTEXT_TOKEN, (instance) => {
+			this._modalService = instance;
+		});
 	}
 
 	async #getPackages() {
@@ -127,6 +134,19 @@ export class UmbPackagesCreatedOverviewElement extends UmbLitElement {
 
 	async #deletePackage(p: PackageDefinitionModel) {
 		if (!p.key) return;
+		const modalHandler = this._modalService?.confirm({
+			color: 'danger',
+			headline: `Remove ${p.name}?`,
+			content: 'Are you sure you want to delete this package',
+			confirmLabel: 'Delete',
+		});
+
+		const deleteConfirmed = await modalHandler?.onClose().then(({ confirmed }: any) => {
+			return confirmed;
+		});
+
+		if (!deleteConfirmed == true) return;
+
 		const { error } = await tryExecuteAndNotify(this, PackageResource.deletePackageCreatedByKey({ key: p.key }));
 		if (error) return;
 		const index = this._createdPackages.findIndex((x) => x.key === p.key);
